+++ G03_CENG431_Final/src/fileio/repository/UpdateData.java
@@ -6,17 +6,17 @@ import java.util.Hashtable;
 public class UpdateData {
 	
 	public String name;
-	public Dictionary<String, Float> values;
+	public Dictionary<String, Double> values;
 	public UpdateData(String name){
 		this.name = name;
-		values = new Hashtable<String,Float>();
+		values = new Hashtable<String,Double>();
 	}
 	
-	public void addKeyValue(String name, Float value){
+	public void addKeyValue(String name, Double value){
 		this.values.put(name, value);
 	}
 	
-	protected  Dictionary<String, Float> getValues(){
+	protected  Dictionary<String, Double> getValues(){
 		return this.values;
 	}
 }
