+++ G03_CENG431_Final/src/model/Coin.java
@@ -4,7 +4,7 @@ import java.util.Dictionary;
 
 public class Coin extends Currency{
 	
-	public Coin(String id, Dictionary<String, Float> value, String name) {
+	public Coin(String id, Dictionary<String, Double> value, String name) {
 		super(id, value, name);
 	}
