+++ G03_CENG431_Final/src/model/ICurrency.java
@@ -3,6 +3,6 @@ package model;
 import java.util.Dictionary;
 
 public interface ICurrency{
-	public Dictionary<String,Float> getValue();
-	public Dictionary<String,Float> getOldValue();
+	public Dictionary<String,Double> getValue();
+	public Dictionary<String,Double> getOldValue();
 }
